add check to test

+++ reduction/test/test_reduction.py
@@ -52,7 +52,8 @@ def test_reduce_workflow():
     template_path = "data/template.xml"
     output_dir = "/tmp"
     reduced_path = os.path.join(output_dir, "REFL_198409_combined_data_auto.txt")
-    os.remove(reduced_path)
+    if os.path.exists(reduced_path):
+        os.remove(reduced_path)
 
     for i in range(198409, 198417):
         ws = mtd_api.Load("REF_L_%s" % i)
@@ -80,7 +81,8 @@ def test_reduce_workflow_201282():
     template_path = "data/template_201282.xml"
     output_dir = "/tmp"
     reduced_path = os.path.join(output_dir, "REFL_201282_combined_data_auto.txt")
-    os.remove(reduced_path)
+    if os.path.exists(reduced_path):
+        os.remove(reduced_path)
 
     for i in range(201282, 201289):
         ws = mtd_api.Load("REF_L_%s" % i)
@@ -108,7 +110,8 @@ def test_background_subtraction():
     template_path = "data/template_short_nobck.xml"
     output_dir = "/tmp"
     reduced_path = os.path.join(output_dir, "REFL_198382_combined_data_auto.txt")
-    os.remove(reduced_path)
+    if os.path.exists(reduced_path):
+        os.remove(reduced_path)
 
     for i in range(198388, 198390):
         ws = mtd_api.Load("REF_L_%s" % i)
